fix Go 1.17 compilation
