+++ row_buffer.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/segmentio/parquet-go/deprecated"
 	"github.com/segmentio/parquet-go/encoding"
-	"github.com/segmentio/parquet-go/internal/unsafecast"
 )
 
 // RowBuffer is an implementation of the RowGroup interface which stores parquet
@@ -190,42 +189,6 @@ func (buf *RowBuffer[T]) reserve(n int) {
 		copy(newRows, buf.rows)
 		buf.rows = newRows
 	}
-}
-
-type rowAllocator struct{ buffer []byte }
-
-func (a *rowAllocator) reset() {
-	a.buffer = a.buffer[:0]
-}
-
-func (a *rowAllocator) capture(row Row) {
-	for i, v := range row {
-		switch kind := v.Kind(); kind {
-		case ByteArray, FixedLenByteArray:
-			row[i].ptr = unsafecast.AddressOfBytes(a.copyBytes(v.byteArray()))
-		}
-	}
-}
-
-func (a *rowAllocator) copyBytes(v []byte) []byte {
-	if free := cap(a.buffer) - len(a.buffer); free < len(v) {
-		newCap := 2 * cap(a.buffer)
-		if newCap == 0 {
-			newCap = defaultReadBufferSize
-		}
-		for newCap < len(v) {
-			newCap *= 2
-		}
-		a.buffer = make([]byte, 0, newCap)
-	}
-
-	i := len(a.buffer)
-	j := len(a.buffer) + len(v)
-	a.buffer = a.buffer[:j]
-
-	b := a.buffer[i:j:j]
-	copy(b, v)
-	return b
 }
 
 type rowBufferColumnChunk struct{ page rowBufferPage }
